Move Settings out of it's own namespace into the app namespace It is only one file now
