+++ app/Settings.php
@@ -22,7 +22,7 @@
  * @copyright  2016 Tony Murray
  * @license    @license http://opensource.org/licenses/GPL-3.0 GNU Public License v3 or later
  */
-namespace App\Settings;
+namespace App;
 
 
 use App\Models\DbConfig;
